requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
elasticsearch==8.11.0
redis==5.0.1
langchain==0.0.340
sentence-transformers==2.2.2
openai==1.3.5
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
markdown==3.5.1
pytest==7.4.3
black==23.11.0
isort==5.12.0
pre-commit==3.5.0
anthropic==0.5.0
